+++ src/main/java/com/linweiyun/genshin/MinegenshinClient.java
@@ -5,6 +5,7 @@ import com.linweiyun.genshin.render.render.entity.FieldTalismanSpiritRender;
 import com.linweiyun.genshin.render.gui.screens.ScreenArtifaceInfo;
 import com.linweiyun.genshin.core.system.registry.register.ModEntities;
 import com.linweiyun.genshin.core.system.registry.register.ModMenus;
+import net.minecraft.client.renderer.entity.LivingEntityRenderer;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.ModContainer;
@@ -47,5 +48,7 @@ public class MinegenshinClient {
                 ModEntities.SLIME_CYRO.get(), context -> new GeoEntityRenderer<>(context, ModEntities.SLIME_CYRO.get())
         );
     }
+
+
 
 }